snooper/pyproject.toml
[project]
name = "snooper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "pillow>=11.1.0",
    "pynput>=1.7.7",
    "pyqt5>=5.15.11",
    "requests>=2.32.3",
]
